+++ application/modules/bpm/controllers/case_manager.php
@@ -25,7 +25,7 @@ class Case_manager extends MX_Controller {
         $this->module_path = 'application/modules/bpm/';
         //----LOAD LANGUAGE
         $this->lang->load('library', $this->config->item('language'));
-        $this->idu = (float) $this->session->userdata('iduser');
+        $this->idu = (int) $this->session->userdata('iduser');
         $this->base_url = base_url();
         $this->module_url = base_url() . 'bpm/';
     }
