+++ application/modules/user/controllers/rbac_admin.php
@@ -14,7 +14,7 @@ class rbac_admin extends MX_Controller {
         $this->user->authorize();
         //----LOAD LANGUAGE
         //$this->lang->load('library', $this->config->item('language'));
-        $this->idu = (float) $this->session->userdata('iduser');
+        $this->idu = (int) $this->session->userdata('iduser');
     }
 
     function repository($action) {
